versiones 12403-12423 Desde SVN. Unificación Coresip Ulises-ETM

+++ ulises-ta/Tools/pjproject/pjlib-util/src/pjlib-util/base64.c
@@ -114,7 +114,7 @@ PJ_DEF(pj_status_t) pj_base64_encode(const pj_uint8_t *input, int in_len,
 	po += 4;
     }
 
-    *out_len = po - output;
+    *out_len = (int)(po - output);
     return PJ_SUCCESS;
 }
 
@@ -123,7 +123,7 @@ PJ_DEF(pj_status_t) pj_base64_decode(const pj_str_t *input,
 				     pj_uint8_t *out, int *out_len)
 {
     const char *buf = input->ptr;
-    int len = input->slen;
+    int len = (int)input->slen;
     int i, j;
     int c1, c2, c3, c4;
